refactor: move outputdata creation into DefaultOutputDataCreator
